STK snapshot 2020-08-06

+++ packages/stk/stk_transfer/stk_transfer/ReducedDependencyGeometricTransfer.hpp
@@ -176,12 +176,12 @@ void create_offset_and_num_key(const std::vector<int> & uniqueProcVec,
 {
   for(unsigned iproc = 0; iproc < uniqueProcVec.size(); ++iproc)
   {
-    const unsigned procid = uniqueProcVec[iproc];
+    const int procid = uniqueProcVec[iproc];
     int offset = -1;
     int numKeys = 0;
     for(int jj = 0; jj < (int) entity_key_proc.size(); ++jj)
     {
-      if(entity_key_proc[jj].proc() == procid)
+      if(static_cast<int>(entity_key_proc[jj].proc()) == procid)
       {
         if(offset < 0)
           offset = jj;
